+++ multiqc/plots/linegraph.py
@@ -6,7 +6,6 @@ from __future__ import print_function
 from collections import OrderedDict
 import base64
 import io
-import json
 import logging
 import os
 import random
@@ -25,7 +24,7 @@ except Exception as e:
     # The lack of the library will be handled when plots are attempted
     print("##### ERROR! MatPlotLib library could not be loaded!    #####", file=sys.stderr)
     print("##### Flat plots will instead be plotted as interactive #####", file=sys.stderr)
-    logger.exception(e)
+    print(e)
 
 letters = 'abcdefghijklmnopqrstuvwxyz'
 
@@ -142,9 +141,14 @@ def highcharts_linegraph (plotdata, pconfig=None):
     if pconfig is None:
         pconfig = {}
 
-    # Build the HTML for the page
+    # Get the plot ID
     if pconfig.get('id') is None:
         pconfig['id'] = 'mqc_hcplot_'+''.join(random.sample(letters, 10))
+
+    # Sanitise plot ID and check for duplicates
+    pconfig['id'] = report.save_htmlid(pconfig['id'])
+
+    # Build the HTML for the page
     html = '<div class="mqc_hcplot_plotgroup">'
 
     # Buttons to cycle through different datasets
@@ -170,16 +174,13 @@ def highcharts_linegraph (plotdata, pconfig=None):
     # The plot div
     html += '<div class="hc-plot-wrapper"><div id="{id}" class="hc-plot not_rendered hc-line-plot"><small>loading..</small></div></div></div> \n'.format(id=pconfig['id'])
 
-    # Javascript with data dump
-    html += '<script type="text/javascript"> \n\
-        mqc_plots["{id}"] = {{ \n\
-            "plot_type": "xy_line", \n\
-            "datasets": {d}, \n\
-            "config": {c} \n\
-        }} \n\
-    </script>'.format(id=pconfig['id'], d=json.dumps(plotdata), c=json.dumps(pconfig));
-
     report.num_hc_plots += 1
+
+    report.plot_data[pconfig['id']] = {
+        'plot_type': "xy_line",
+        'datasets': plotdata,
+        'config': pconfig
+    }
 
     return html
 
@@ -196,6 +197,10 @@ def matplotlib_linegraph (plotdata, pconfig=None):
     # Plot group ID
     if pconfig.get('id') is None:
         pconfig['id'] = 'mqc_mplplot_'+''.join(random.sample(letters, 10))
+
+    # Sanitise plot ID and check for duplicates
+    pconfig['id'] = report.save_htmlid(pconfig['id'])
+
     # Individual plot IDs
     pids = []
     for k in range(len(plotdata)):
@@ -204,7 +209,7 @@ def matplotlib_linegraph (plotdata, pconfig=None):
         except:
             name = k+1
         pid = 'mqc_{}_{}'.format(pconfig['id'], name)
-        pid = "".join([c for c in pid if c.isalpha() or c.isdigit() or c == '_' or c == '-'])
+        pid = report.save_htmlid(pid)
         pids.append(pid)
 
     html = '<p class="text-info"><small><span class="glyphicon glyphicon-picture" aria-hidden="true"></span> ' + \
@@ -254,6 +259,7 @@ def matplotlib_linegraph (plotdata, pconfig=None):
                         fdata[d['name']][pconfig['categories'][i]] = x
                     except (KeyError, IndexError):
                         fdata[d['name']][str(i)] = x
+
         # Custom tsv output if the x axis varies
         if not sharedcats and config.data_format == 'tsv':
             fout = ''
@@ -370,7 +376,7 @@ def matplotlib_linegraph (plotdata, pconfig=None):
 
         # Tight layout - makes sure that legend fits in and stuff
         if len(pdata) <= 15:
-            lgd = axes.legend(loc='lower center', bbox_to_anchor=(0, -0.22, 1, .102), ncol=5, mode='expand', fontsize=8, frameon=False)
+            axes.legend(loc='lower center', bbox_to_anchor=(0, -0.22, 1, .102), ncol=5, mode='expand', fontsize=8, frameon=False)
             plt.tight_layout(rect=[0,0.08,1,0.92])
         else:
             plt.tight_layout(rect=[0,0,1,0.92])