+++ multiqc/utils/config.py
@@ -6,6 +6,7 @@ config variables to be used across all other modules """
 from __future__ import print_function
 from datetime import datetime
 import inspect
+import collections
 import os
 import pkg_resources
 import random
@@ -51,7 +52,7 @@ with open(searchp_fn) as f:
 
 # Other defaults that can't be set in YAML
 modules_dir = os.path.join(MULTIQC_DIR, 'modules')
-creation_date = datetime.now().strftime("%Y-%m-%d, %H:%m")
+creation_date = datetime.now().strftime("%Y-%m-%d, %H:%M")
 working_dir = os.getcwd()
 analysis_dir = [os.getcwd()]
 output_dir = os.path.realpath(os.getcwd())
@@ -88,7 +89,8 @@ if len(avail_modules) == 0 or len(avail_templates) == 0:
         the installation script (python setup.py install)", file=sys.stderr)
     sys.exit(1)
 
-# Functions to load user config files. These are called by the main MultiQC script.
+##### Functions to load user config files. These are called by the main MultiQC script.
+# Note that config files are loaded in a specific order and values can overwrite each other.
 def mqc_load_userconfig(path=None):
     """ Overwrite config defaults with user config files """
 
@@ -117,22 +119,7 @@ def mqc_load_config(yaml_config):
             with open(yaml_config) as f:
                 new_config = yaml.load(f)
                 logger.debug("Loading config settings from: {}".format(yaml_config))
-                for c, v in new_config.items():
-                    if c == 'sp':
-                        # Merge filename patterns instead of replacing
-                        sp.update(v)
-                        logger.debug("Added to filename patterns: {}".format(v))
-                    elif c == 'extra_fn_clean_exts':
-                        # Prepend to filename cleaning patterns instead of replacing
-                        fn_clean_exts[0:0] = v
-                        logger.debug("Added to filename clean extensions: {}".format(v))
-                    elif c == 'extra_fn_clean_trim':
-                        # Prepend to filename cleaning patterns instead of replacing
-                        fn_clean_trim[0:0] = v
-                        logger.debug("Added to filename clean trimmings: {}".format(v))
-                    else:
-                        logger.debug("New config '{}': {}".format(c, v))
-                        globals()[c] = v
+                mqc_add_config(new_config)
         except (IOError, AttributeError) as e:
             logger.debug("Config error: {}".format(e))
         except yaml.scanner.ScannerError as e:
@@ -141,3 +128,71 @@ def mqc_load_config(yaml_config):
     else:
         logger.debug("No MultiQC config found: {}".format(yaml_config))
 
+def mqc_cl_config(cl_config):
+    for clc_str in cl_config:
+        try:
+            parsed_clc = yaml.load(clc_str)
+            assert(isinstance(parsed_clc, dict))
+        except yaml.scanner.ScannerError as e:
+            logger.error("Could not parse command line config: {}\n{}".format(clc_str, e))
+        except AssertionError:
+            logger.error("Could not parse command line config: {}".format(clc_str))
+        else:
+            mqc_add_config(parsed_clc)
+
+def mqc_add_config(conf):
+    """ Add to the global config with given MultiQC config dict """
+    global fn_clean_exts, fn_clean_trim
+    for c, v in conf.items():
+        if c == 'sp':
+            # Merge filename patterns instead of replacing
+            sp.update(v)
+            logger.debug("Added to filename patterns: {}".format(v))
+        elif c == 'extra_fn_clean_exts':
+            # Prepend to filename cleaning patterns instead of replacing
+            fn_clean_exts[0:0] = v
+            logger.debug("Added to filename clean extensions: {}".format(v))
+        elif c == 'extra_fn_clean_trim':
+            # Prepend to filename cleaning patterns instead of replacing
+            fn_clean_trim[0:0] = v
+            logger.debug("Added to filename clean trimmings: {}".format(v))
+        else:
+            logger.debug("New config '{}': {}".format(c, v))
+            update_dict(globals(), {c: v})
+
+#### Function to load file containinga list of alternative sample-name swaps
+# Essentially a fancy way of loading stuff into the sample_names_rename config var
+# As such, can also be done directly using a config file
+def load_sample_names(snames_file):
+    global sample_names_rename_buttons, sample_names_rename
+    num_cols = None
+    try:
+        with open(snames_file) as f:
+            logger.debug("Loading sample renaming config settings from: {}".format(snames_file))
+            for l in f:
+                s = l.strip().split("\t")
+                if len(s) > 1:
+                    # Check that we have consistent numbers of columns
+                    if num_cols is None:
+                        num_cols = len(s)
+                    elif num_cols != len(s):
+                        logger.warn("Inconsistent number of columns found in sample names file (skipping line): '{}'".format(l.strip()))
+                    else:
+                        # Parse the line
+                        if len(sample_names_rename_buttons) == 0:
+                            sample_names_rename_buttons = s
+                        else:
+                            sample_names_rename.append(s)
+    except (IOError, AttributeError) as e:
+        logger.error("Error loading sample names file: {}".format(e))
+    logger.debug("Found {} sample renaming patterns".format(len(sample_names_rename_buttons)))
+
+def update_dict(d, u):
+    """ Recursively updates nested dict d from nested dict u
+    """
+    for key, val in u.items():
+        if isinstance(val, collections.Mapping):
+            d[key] = update_dict(d.get(key, {}), val)
+        else:
+            d[key] = u[key]
+    return d