+++ multiqc/utils/report.py
@@ -10,6 +10,7 @@ import click
 import fnmatch
 import io
 import json
+import lzstring
 import mimetypes
 import os
 import re
@@ -32,6 +33,8 @@ general_stats_data = list()
 general_stats_headers = list()
 general_stats_html = ''
 data_sources = defaultdict(lambda:defaultdict(lambda:defaultdict()))
+plot_data = dict()
+html_ids = list()
 num_hc_plots = 0
 num_mpl_plots = 0
 saved_raw_data = dict()
@@ -230,4 +233,38 @@ def data_sources_tofile ():
             body = '\n'.join(["\t".join(l) for l in lines])
             print( body.encode('utf-8', 'ignore').decode('utf-8'), file=f)
 
+def save_htmlid(html_id):
+    """ Take a HTML ID, sanitise for HTML, check for duplicates and save.
+    Returns sanitised, unique ID """
+    global html_ids
+
+    # Trailing whitespace
+    html_id = html_id.strip()
+
+    # Must begin with a letter
+    if re.match(r'^[a-zA-Z]', html_id) is None:
+        html_id = 'mqc-{}'.format(html_id)
+
+    # Replace illegal characters
+    html_id = re.sub('[^\w:.-]+', '-', html_id)
+
+    # Check for duplicates
+    i = 1
+    html_id_base = html_id
+    while html_id in html_ids:
+        html_id = '{}-{}'.format(html_id_base, i)
+        i += 1
+
+    # Remember and return
+    html_ids.append(html_id)
+    return html_id
+
+
+def compress_json(data):
+    """ Take a Python data object. Convert to JSON and compress using lzstring """
+    json_string = json.dumps(data).encode('utf-8', 'ignore').decode('utf-8')
+    # JSON.parse() doesn't handle `NaN`, but it does handle `null`.
+    json_string = json_string.replace('NaN', 'null');
+    x = lzstring.LZString()
+    return x.compressToBase64(json_string)